Function dicediele::terminal::t_roll_request

pub fn t_roll_request() -> String

Creates a loop where the user will be prompted to specify which dice to roll, be presented with the outcomes, and asked if they would like to repeat.

After specifying which dice to roll, the user will be given the value of each roll as well as summary values such as the highest rolled value and total value. They will then be asked if they would like to roll again and can either accept to repeat the loop or decline to exit it. Alternatively, they can exit early by tying menu.

This function is a WiP and could be split into smaller functions handling sections of the loop to hand more control to main.rs Terminal prompt for rolling input. Returns user input as a string.

Asks the user to specify what/how many dice to roll. Trims whitespace from the input before returning.

Examples

use dicediele::terminal_io::*;
 
// Ask the user which dice they would like to roll and store their input
let dice_inputs: String = terminal_io::t_roll_request();