Function dicediele::terminal_io::t_roll_request

fn t_roll_request() -> [u32; 2]

Terminal prompt for rolling input. Returns processed user input as an array.

Asks the user to specify what/how many dice to roll. Expects a value in the format input, where N is the number of X-sided dice to roll. Also parses the input and returns a [u32;2] array containing [N, X]. Alternatively, returns [0,0] if the user types cancel.

Examples

use dicediele::terminal_io::*;
use dicediele::dice::*;
 
// Ask the user which dice they would like to roll and store their input
let dice_inputs: [u32; 2] = terminal_io::t_roll_request();
 
// Pass the user input to a function which rolls input
let rolls: Vec<u32> = dice::roll_input(dice_inputs[0], dice_inputs[1]);
 
// Print the results of each die as well as the total on separate lines
for result in &rolls {
    println!("You rolled a {}.", result);
}
let sum: u32 = rolls.iter().sum();
println!("Your total is {}.", sum);